Implemented a browse action in TIP_Class
To avoid killing myself in this operation (really complex to do
in an elegant way) I left the logic complexity in the database.

The browse action works by inspecting a _view_ provided by the
configuration option ("summary_view") and the actionBrowse()
method temporary switches the current data source ($this->data)
to this view right before running the browse template.