[TIP_Mysql] Modified alias management in joins
The "alias" property approach is wrong: it is not possible to
connect the same table more than once.

'joins'           => array(
    'slave-table' => array(
        'alias'   => 'slave1',
        'master'  => 'id1',
        'slave'   => 'id'
    ),
    'slave-table' => array(
        'alias'   => 'slave2',
        'master'  => 'id2',
        'slave'   => 'id'
    )
)

This doesn't work because the original table is the KEY of the array.
This is the new (correct) approach:
                                                                                   ),
'joins'           => array(
    'slave1'      => array(
        'path'    => 'slave-table',
        'master'  => 'id1',
        'slave'   => 'id'
    ),
    'slave2'      => array(
        'path'    => 'slave-table',
        'master'  => 'id2',
        'slave'   => 'id'
    )
)

The alias must be used as KEY, because this is the way SQL works.
Now slave-table could be connected as many times as needed
(as long as a new alias name is specified).