[TIP_Form] Implemented "placeholder" flag
By tagging a field as "placeholder", when the default value is
undefined (typically while adding a new row) an INSERT query
will be performed in the table specified in "widget_args" and
the last inserted id will be used to fill the field value.

Now the query is "INSERT INTO widget_args () VALUES ()". If
needed, the widget_args meaning could be extended to provide
default values to the placehold table.