[TIP_Form] Implemented custom filter for "set" and "enum" widgets
By default, set and enum widgets (automatically generated on SET
and ENUM fields) give a widget exposing all the availabe options
(got from the field info). Now a custom filter can be specified as
"widget_args" to allow or deny the default expose of values.

"widget_args" can contain an allow rule in the form
"allow(choice1 choice2 ...)", in which case only the specified
list of space separated choices will be used, or a deny rule in the
form "deny(choice1 choice2 ...)", in which case the specified
choices will be stripped from the available choices. For instance,
specifying "widget_args=deny(firm person)" will strip "firm" and
"person" from the exposed choices.

The implementation also takes into account the default selection,
that must not be filtered. This means in the above example the
"firm" and "person" choices can be selected by explicitely set
them (providing a proper GET or via application).