[TIP_Content] Search improvements
Allows to specify an array in the "search_field" property. Also,
the array can be associative (in the form "alias" => "field") in
the same way field can be specified in joins. This makes possible
to specify the following "search_field" value:

array('propertydetail' => '@`property-detail`.`detail`')

In this case, `property-detail`.`detail` will be physically used
inside the query and "propertydetail" will be considered while
highlighting the tags.