[TIP_Form] Implemented "notify_done" property
This boolean property let the form module to work as before
if set to true (the default) and disallows the notification
to the user on succesful actions if set to false.

This allows, for instance, to show the default page (shown
only if the page content is empty) after a succesful login.