[TIP_Form] Force defaults for "referer" and "follower"
Using TIP::arrayDefault() was treating a null value as default,
while "referer" and "follower" should be resolved to the referer URI.
By calling isset() || ... instead of TIP::arrayDefault() will
set the default value also if null is set in the configuration option.

This enables to cancel the value set programmaticaly by the module:
setting explicitely a null value in the configuration option will
restore the default behaviour.