[TIP_User] The login action now needs "none" privilege
Previously the login was needing "untrusted" privilege, as the
"add" action, disallowing the separation of login and registration.
A typical example is a locked context, where the new users are
registered by an administrator.

With a default "none" privilege for login actions, the anonymous
privilege of the user module can be set to TIP_PRIVILEGE_NONE to
deal with such common situation.