[TIP_User] Redistribuited privilege actions
The user module needs to fulfill locked/unlocked context (that is
a site where anonymous users can register themselves or a site
where the user registration are managed by an administrator).

To be able to do this with a common framework, the user module
actions are distribuited in this way on a per-level basis:

- TIP_PRIVILEGE_NONE: can request a log-in
- TIP_PRIVILEGE_UNTRUSTED: can register himself if anonymous or
  edit/delete his own profile and log-out if authenticated
- TIP_PRIVILEGE_TRUSTED: can register another user
- TIP_PRIVILEGE_ADMIN: can view and delete other user profiles
- TIP_PRIVILEGE_MANAGER: can modify every user profile

In this way, to get an open site (free registration), you can
set the user module privileges in this way:

'user'                    => array(
    'type'                => array('module', 'content', 'user'),
    'anonymous_privilege' => TIP_PRIVILEGE_UNTRUSTED,
    'default_privilege'   => TIP_PRIVILEGE_UNTRUSTED,
    ...

Instead, if you need a locked context, where an administrator
is the only responsible of registering new users:

'user'                    => array(
    'type'                => array('module', 'content', 'user'),
    'anonymous_privilege' => TIP_PRIVILEGE_NONE,
    'default_privilege'   => TIP_PRIVILEGE_UNTRUSTED,
    ...

The user administrator should have at least a TIP_PRIVILEGE_TRUSTED
level.