[TIP_User] Avoid autologin on request from a registered user
While previously the autologin worked on every actionAdd() request
on the user module, if the site is locked (that is the new users
are added by an administrator) this approach doesn't work because
when the administrator adds a new user there's a switch to the
newly user.

This patch performs the autologin only if the request comes from
an anonymous user.