[TIP_Form] Implemented a prerendered system
Now a serie of callbacks could be called before the rendering
process. This allows to do custom operations based on the freeze
state. For instance, the hierarchy widget can be populated with
the whole table content (as before) or with only the needed row
if its state is frozen (not possible before).