[TIP_Class] Freeze the "class_field" if not null
While previously an add form allowed to change the "class_field"
element after it was yet selected, this has some drawback: mainly
the form requires a submission to change the child form and it is
unexpectedly validated. Now the "class_field" is always frozen
once selected: to change it you must cancel the current operation
and redo it selecting another class.