[TIP_Form] Using "info" instead of "comment"
Additional field details (to be used as a kind of "tooltip" while
rendering the forms) now uses the new get/setInfo() API of
HTML_QuickForm_Element. This because the get/setComment() API is
used to add a comment after the element and will be used to pass
arguments to the javascript logic, if any.