[TIP_Form] Implemented the TIP_FORM_RENDER_REDIRECT target
Sometime is useful to redirect to a summary page after a form action
instead of reshow the same frozen data. This new target allows to do
this, redirecting to the "follower" URI. The form action result is
notified by TIP_Application::tagPage() at the beginning of the page:
the "notify" session variable is used to communicate the data.