[TIP_Application] Handle of string 'id'
The id is not always an integer value: sometime is useful to use
string values. Anyway, to avoid SQL injection, this field must be
treated really careful. This commit inspects the $cfg global
configuration array to check if the current module as an explicit
'id_type' property different from 'integer' and uses it as type
casting to $_GET['id'] and $_POST['id']. If not specified, it
defaults to 'integer', as previously forced by TIP_Application.