[TIP] Removed trailing slash from URIs
Modified the buildURI() family functions to avoid appending a slash
at the end of every URI. Logically speaking, a URI with a trailing
slash is a directory (kind of) while the URI usually used are
a single file. This new implementation also let internal URIs
(page.html#section) to work properly.