[TIP_Content] Removed bug in internal link rendering
Previously the destination of an internal link was appended to a
base URL. This gives some problems, such as appending the id
to an URL ending with a letter.

The new implementation simply leverages on the proper URL rendering
of Text_Wiki, that is using a %s as a tag.