[TIP_XML] Improved XML data engine
Well defined the XML data engine: added the "fields_xpath" property
(containing the mapping between field id and xpaths) and forced its
definition. The manual mapping of every field allows a greater (by
far) flexibility at the cost of more complexity.

If the "path" property begins with "http://", the path is read as
is (that is, the file is fetched from the web) instead of being
parsed by TIP::buildDataPath() as needed by local files.

The new approach is to always parse the whole XML file into an array
of rows on the first call and run the queries on the cached rows.
This enables more complex filtering by using array functions and
leaving intact, for instance, the calculated parent field or the
eventually autogenerated primary key.

Actually, simple SQL queries are recognized if in the form:
 WHERE field = value [LIMIT length[,offset]]