[TIP_Application] Get the "id_type" property from the module instance
To force a type casting on the "id" it is necessary to know its type:
this is a property stored in the module itsself. Previously, this
property was got from the global $cfg configuration array but this
approach has some pitfall, first of all it does not catch "id_type"
default values programmatically set by the module.

Now the type casting is posticipated up the module is loaded, so the
black magic is no longer needed.