[TIP_Data_Engine] Added $domain to preparedName() API
To deal with really different data engines, it is better to let the
engine choose the way to specify the domain of a name. The domain
is, for instance, the table for SQL field names (the "path" property
of TIP_Data), but in the TIP_XML engine this does not make sense.
In such situation, the new API allows to ignore the domain feature
directly from preparedName().