[TIP_Data] Using the domain feature of preparedName()
Updated TIP_Data to use the domain feature instead of implying
the domain must be prepared and prepended with a dot. Also,
removed the hardcoded '@' to enable raw format, as this should
be an engine dependent feature.

THIS IS AN API BREAKAGE, as the '@' char prepended to fields
is no more considered a special case.