[PEAR] Enabled inline YouTube movies
Hackish commit to Url.php that enables the embedding of an
hardcoded xhtml <div> if a movie matching a regex is found
(by default, the regex is "|http://www.youtube.com/v/|").

This is a somewhat naive implementation but having anyway
to rewrite the whole wiki engine there is no motivations
to be elegant.