[TIP_Content] Enhanced chronology to support a custom id
Previously, the TIP_Chronology id of a specific TIP_Content was
the content id + "_chronology". Now the tagChronology() use
the first TIP_Chronology type that has the current content module
as "master" option.

The inability to implement more than one chronology for a single
content it is still an open issue.