[TIP_Form] Corrected bug in name mangling
Composite widgets (hierarchy and lookup) was constructing a fallback
parent module id by using the 'form' prefix because TIP_Form was
lacking an explicit id. Forced the form id to be the same as the
master module (as it was previously intended).

Also, while before the parent module of such widgets, if not
explicitely defined, was searched by inspecting the $cfg global
array for a module with the form module id as "master" option,
now it is also required that the the module is of TIP_Hierarchy type.