[TIP_Boss1] First (working) implementation
Implemented a module that leverages the Yahoo! BOSS API (v.1)
to do a local search on a website.

This first implementation should be enhanced: although it is
able to use the default pager provided by TIP_Content, the
XML engine does not support summary fields, loosing the value
of count, start, totalhits or deephits, needed to know
if there is a previous or next page.