Corrected thumbnail managemen in TIP_Picasa2
Previously, the thumbnails were catalogued with a selector based on the
width of the image, expecting for example a standard width of 288 pixels.

This is not the case, because when the photo is in portrait the width is
unknown while the height is 288 pixels.

To overcome this issue, the new approach always takes the third
thumbnails, expecting it to be a 288 width or height image.