Corrected bug in TOC rendering
The $min property value in Text_Wiki_Render_Xhtml_Toc is not properly
handled: the renderer does not care when an <h1> element is present and
generated invalid Xhtml putting the h1 text in an <li> element not
enclosed by an <ul>.

Modified to properly handling this case. A change to the parser has been
needed to skip the generation of the header text.